Update Dockerfile configuration
This update is mainly targeted at enhancing the Dockerfile configuration. It involved moving the working directory to the '/app' directory from '.', replacing the link copy operation of the package.json file with a normal copy, and also including yarn.lock file. The update also improves the command to copy contents from the build image to the production image by changing the source and destination directories.